std: move locks to `sys` on xous
