+++ library/std/src/sys/locks/condvar/xous.rs
@@ -1,6 +1,6 @@
-use super::mutex::Mutex;
 use crate::os::xous::ffi::{blocking_scalar, scalar};
 use crate::os::xous::services::{ticktimer_server, TicktimerScalar};
+use crate::sys::locks::Mutex;
 use crate::time::Duration;
 use core::sync::atomic::{AtomicUsize, Ordering};
