--- library/std/src/sys/pal/xous/locks/mod.rs
@@ -1,7 +1,0 @@
-mod condvar;
-mod mutex;
-mod rwlock;
-
-pub use condvar::*;
-pub use mutex::*;
-pub use rwlock::*;
