+++ library/std/src/sys/pal/xous/mod.rs
@@ -9,7 +9,6 @@ pub mod env;
 pub mod fs;
 #[path = "../unsupported/io.rs"]
 pub mod io;
-pub mod locks;
 pub mod net;
 pub mod os;
 #[path = "../unsupported/pipe.rs"]
